Uploaded my RELAY EXEC
This exec is meant to communicate to the RELAY service maintained by Moshix on the HNET.
There is one thing to consider: on my CMS I had to prepend "CP " before the SMSG command and this causes the arguments to be truncated to 8 characters. You might want to try the "CP " less version by uncommenting that line.